requirements.txt
absl-py==0.8.0
astor==0.8.0
click==7.1.1
cycler==0.10.0
dash==1.9.1
dash-bootstrap-components==0.8.3
dash-core-components==1.8.1
dash-html-components==1.0.2
dash-renderer==1.2.4
dash-table==4.6.1
Flask==1.1.1
Flask-Compress==1.4.0
future==0.18.2
gast==0.3.2
google-pasta==0.1.7
grpcio==1.23.0
h5py==2.10.0
itsdangerous==1.1.0
Jinja2==2.11.1
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
kiwisolver==1.1.0
Markdown==3.1.1
MarkupSafe==1.1.1
matplotlib==3.1.1
numpy==1.14.5
opencv-python==4.1.1.26
Pillow==6.1.0
plotly==4.5.3
protobuf==3.9.1
pyparsing==2.4.2
python-dateutil==2.8.0
retrying==1.3.3
six==1.12.0
tensorboard==1.14.0
tensorflow==1.14.0
tensorflow-estimator==1.14.0
termcolor==1.1.0
Werkzeug==0.16.0
wrapt==1.11.2
